Introduce ResultViewer component and use it in the results table
The component style is based on the question and answers component.
Specific answers can highlighted. We use it to display the last selection
the user made. For that we need to extend the ResultType with a selections
property. This property will be updated every time the user submits an answer.

Extend result component tests and update documentation